Format body and change hovered grid squares
A flexbox was used to display the button and grid in a more orderly
fashion. Some formatting was applied for a cleaner display while also
bumping down the grid on the page so the user did not accidentally draw
when using the popup window.

Grid squares now change their background color when they have been
hovered over to track the mouse movement and allow the user to draw.