Adjust layout of page
Made the page into a few flex boxes to format the page layout so the
button for changing the grid size is on the side of the sketchpad. Gave
space at the top of the page in the form of a header so the popup does
not cover the sketchpad.